committing

+++ src/utils/cloudinary.js
@@ -7,20 +7,32 @@ cloudinary.config({
   api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
-const uploadOnCloudinary = async (localFilePath) =>{
-    try {
-        if(!localFilePath) return null;
+const uploadOnCloudinary = async (localFilePath) => {
+  try {
+    if (!localFilePath) return null;
 
-        const response = await cloudinary.v2.uploader.upload(localFilePath,{
-            resource_type: "auto"
-        })
-        console.log("File has been uploaded successfully ", response.url);
-        return response;
-    } catch (error) {
-        fs.unlinkSync(localFilePath)
-        return null;
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto"
+    });
+
+    console.log("Uploaded to Cloudinary:", response.url);
+
+    
+    fs.unlinkSync(localFilePath);
+
+    return response;
+
+  } catch (error) {
+    
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
+
+    console.error("Cloudinary upload error:", error);
+    return null;
+  }
 }
+
 
 
 // cloudinary.v2.uploader.upload("https://logos-download.com/wp-content/uploads/2016/03/Wikipedia_logo_logotype_emblem.png",
